+++ godfather/roles/mafia/consigliere.py
