+++ godfather/roles/mafia/goon.py
