+++ godfather/roles/neutral/serial_killer.py
