+++ godfather/roles/neutral/survivor.py
