+++ godfather/roles/town/vigilante.py
